Return telemetry for player vehicle in rf2mapper.
Previously, this code returned the vehicle at index zero in the telemetry and scoring arrays. However, to correctly find the player vehicle we need to scan the entire array to find the vehicle that is marked as player-controlled.

This has been tested with LMU, with RPM lights and sound-based haptics working after these changes.